update: move showcase content to single component
