+++ src/modules/utils/components/showCase/index.js
@@ -1,0 +1,4 @@
+import { ShowCase } from './ShowCase';
+if (!window.customElements.get(ShowCase.tag)) {
+	window.customElements.define(ShowCase.tag, ShowCase);
+}
