+++ test/modules/header/components/Header.test.js
@@ -1,8 +1,7 @@
 /* eslint-disable no-undef */
-import { render } from 'lit-html';
 import { Header } from '../../../../src/modules/header/components/Header';
 import { sidePanelReducer } from '../../../../src/modules/menue/store/sidePanel.reducer';
-import { mapReducer } from '../../../../src/modules/map/store/olMap.reducer';
+import { modalReducer } from '../../../../src/modules/modal/store/modal.reducer';
 import { TestUtils } from '../../../test-utils.js';
 import { $injector } from '../../../../src/injection';
 import { OlCoordinateService } from '../../../../src/services/OlCoordinateService';
@@ -20,9 +19,10 @@ describe('Header', () => {
 		const state = {
 			sidePanel: {
 				open: false
-			}
+			},
+			modal: { title: false, content: false }
 		};
-		store = TestUtils.setupStoreAndDi(state, { sidePanel: sidePanelReducer });
+		store = TestUtils.setupStoreAndDi(state, { sidePanel: sidePanelReducer, modal: modalReducer });
 		$injector
 			.register('CoordinateService', OlCoordinateService)
 			.registerSingleton('EnvironmentService', { isEmbedded : () => embed });
@@ -83,75 +83,11 @@ describe('Header', () => {
 
 		it('shows a modal window with the showcase', async () => {
 			const element = await setup({ mobile: false });
-			spyOn(element, 'createShowCase');
 
 			element.shadowRoot.querySelector('.ci-logo').click();
 
-			expect(element.createShowCase).toHaveBeenCalled();
+			expect(store.getState().modal.title).toBe('Showcase');
 		});
 	});
-
-	describe('when showcase created', () => {	
-		let store;
-			
-		const extendedSetup = () => {
-			const state = {
-				mobile: false,
-				map: {
-					zoom: 10,
-					pointerPosition: [1288239.2412306187, 6130212.561641981]
-				},
-				sidePanel: {
-					open: false
-				}
-			};
-			store = TestUtils.setupStoreAndDi(state, { map: mapReducer, sidePanel: sidePanelReducer });
-			$injector
-				.register('CoordinateService', OlCoordinateService)
-				.registerSingleton('EnvironmentService', { isEmbedded : () => false });
-			
-			
-			return TestUtils.render(Header.tag);
-		};
-
-		it('adds a div which shows some buttons', async () => {
-			const element = await setup({ mobile: false });
-			const container = document.createElement('div');
-			const showCaseContent = element.createShowCase();
-		
-			render(showCaseContent, container);
-
-			// try different approaches
-			expect(container.querySelectorAll('ba-button').length).toBe(4);
-			expect(container.querySelector('.buttons').childElementCount).toBe(4);
-			expect(container.querySelector('#button0')).toBeTruthy();
-			expect(container.querySelector('#button1')).toBeTruthy();
-			expect(container.querySelector('#button2')).toBeTruthy();
-			expect(container.querySelector('#button3')).toBeTruthy();
-		});
-
-		it('calls the callback, if button0 are clicked', async () => {
-			const element = await extendedSetup();			
-			const container = document.createElement('div');
-			const showCaseContent = element.createShowCase();		
-
-			render(showCaseContent, container);			
-			container.querySelector('#button0').click();
-			
-			expect(store.getState().map.zoom).toBe(13);
-		});
-
-		it('calls the callback, if button1 are clicked', async () => {
-			const element = await extendedSetup();			
-			const container = document.createElement('div');
-			const showCaseContent = element.createShowCase();		
-
-			render(showCaseContent, container);			
-			container.querySelector('#button1').click();
-			
-			expect(store.getState().map.zoom).toBe(11);
-		});
-	});
-
 
 });
